rename miscs to misc
