+++ utils/forest_utils/src/lib.rs
@@ -8,7 +8,7 @@ pub mod encoding;
 pub mod io;
 pub mod json;
 pub mod macros;
-pub mod miscs;
+pub mod misc;
 pub mod monitoring;
 pub mod net;
 pub mod version;
