+++ utils/forest_utils/src/misc/logo.rs
