+++ vm/state_migration/src/lib.rs
@@ -6,7 +6,7 @@ use std::sync::Arc;
 use cid::Cid;
 use forest_networks::{ChainConfig, Height};
 use forest_shim::clock::ChainEpoch;
-use forest_utils::miscs::reveal_five_trees;
+use forest_utils::misc::reveal_five_trees;
 use fvm_ipld_blockstore::Blockstore;
 
 pub(crate) mod common;
